Two more minor warning fixes.
One about an integer conversion, and one about `memcpy` missing because we included `"string.h"` rather than `<string.h>`.

While at it, I also replaced one `#import` with an `#include` and removed another redundant `#include`.

+++ Sources/CgRPC/shim/observers.c
@@ -236,7 +236,7 @@ long cgrpc_observer_recv_status_on_client_get_status(cgrpc_observer_recv_status_
 }
 
 char *cgrpc_observer_recv_status_on_client_copy_status_details(cgrpc_observer_recv_status_on_client *observer) {
-  int length = GRPC_SLICE_LENGTH(observer->server_details);
+  size_t length = GRPC_SLICE_LENGTH(observer->server_details);
   char *str = (char *) malloc(length + 1);
   memcpy(str, GRPC_SLICE_START_PTR(observer->server_details), length);
   str[length] = 0;
